organizing folders WalletScreen & Modals
