--- src/components/Button/TransactionContacts/AddContact/AddContact.styles.ts
@@ -1,15 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-	title: {
-		fontFamily: 'Inter_800ExtraBold',
-		fontStyle: 'normal',
-		fontWeight: '800',
-		fontSize: 24,
-		lineHeight: 29,
-		marginBottom: 32
-	},
-	button: {
-		marginTop: 16
-	}
-});
