+++ src/screens/WalletScreen/Modals/ComingSoonModal/index.tsx
