+++ src/screens/WalletScreen/Modals/ReceiveModal/index.tsx
