+++ src/screens/WalletScreen/Modals/SendModal/components/Card/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		height: 40,
+		flexDirection: 'row',
+		marginBottom: 24
+	},
+	titleContainer: {
+		marginLeft: 16,
+		justifyContent: 'space-between'
+	}
+});
