+++ src/screens/WalletScreen/Modals/SendModal/components/Card/types.ts
@@ -1,0 +1,6 @@
+import { WalletToken } from '@src/model/wallet';
+
+export interface CardProps {
+	token: WalletToken;
+	onSelected?: (token: WalletToken) => void;
+}
