+++ src/screens/WalletScreen/Modals/SendModal/components/ContactItem/types.ts
@@ -1,0 +1,8 @@
+import { ContactItem } from '@src/model/contact';
+
+interface ContactProps {
+	contact: ContactItem;
+	onSelected: (coin: ContactItem) => void;
+}
+
+export type { ContactProps };
