+++ src/screens/WalletScreen/Modals/SendModal/components/GasPriceLine/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		marginBottom: 8,
+		flexDirection: 'row',
+		justifyContent: 'space-between'
+	}
+});
