+++ src/screens/WalletScreen/Modals/SendModal/components/GasPriceLine/types.ts
@@ -1,0 +1,7 @@
+interface GasPriceLineProps {
+	gas: number;
+	label: string;
+	priceUSD: number;
+}
+
+export type { GasPriceLineProps };
