+++ src/screens/WalletScreen/Modals/SendModal/components/Item/styles.ts
@@ -1,0 +1,20 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	itemContainer: {
+		height: 40,
+		flexDirection: 'row',
+		marginBottom: 24,
+		alignItems: 'center'
+	},
+	avatar: {
+		width: 40,
+		height: 40,
+		borderRadius: 20,
+		marginRight: 16
+	},
+	contactTitleContainer: {
+		flex: 1,
+		justifyContent: 'space-between'
+	}
+});
