+++ src/screens/WalletScreen/Modals/SendModal/components/Item/types.ts
@@ -1,0 +1,10 @@
+import { ImageSourcePropType } from 'react-native';
+
+interface ItemProps {
+	firstLine: string;
+	secondLine: string | JSX.Element;
+	imageSource?: ImageSourcePropType;
+	onSelected: () => void;
+}
+
+export type { ItemProps };
