+++ src/screens/WalletScreen/Modals/SendModal/components/index.ts
@@ -1,0 +1,6 @@
+import ContactItem from './ContactItem';
+import Item from './Item';
+import Card from './Card';
+import GasPriceLine from './GasPriceLine';
+
+export { ContactItem, Item, Card, GasPriceLine };
