+++ src/screens/WalletScreen/Modals/SendModal/screens/AddContact/types.ts
@@ -1,0 +1,5 @@
+interface AddContactProps {
+	onContactAdded: () => void;
+}
+
+export type { AddContactProps };
