+++ src/screens/WalletScreen/Modals/SendModal/screens/TransactionContacts/NoContactsYet/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 1,
+		alignItems: 'center',
+		paddingHorizontal: 24
+	},
+	image: {
+		width: 147,
+		height: 137,
+		marginBottom: 32
+	}
+});
