+++ src/screens/WalletScreen/Modals/SendModal/screens/TransactionContacts/styles.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 1,
+		paddingHorizontal: 24
+	}
+});
