+++ src/screens/WalletScreen/Modals/SendModal/screens/TransactionContacts/types.ts
@@ -1,0 +1,7 @@
+import { UserProps } from '../../types';
+
+interface TransactionContactsProps {
+	onSelected: (item: UserProps) => void;
+}
+
+export type { TransactionContactsProps };
