+++ src/screens/WalletScreen/Modals/SendModal/screens/TransactionSelectFunds/types.ts
@@ -1,0 +1,9 @@
+import { WalletToken } from '@src/model/wallet';
+import { UserProps } from '../../types';
+
+interface TransactionSelectFundsProps {
+	user: UserProps;
+	onSelected: (token: WalletToken) => void;
+}
+
+export type { TransactionSelectFundsProps };
