+++ src/screens/WalletScreen/Modals/SendModal/screens/TransactionTransfer/types.ts
@@ -1,0 +1,10 @@
+import { WalletToken } from '@models/wallet';
+import { UserProps } from '../../types';
+import { ResultProps } from '../../../../WalletScreen.types';
+
+export interface TransactionTransferProps {
+	user: UserProps;
+	token: WalletToken;
+	onDismiss: () => void;
+	sentSuccessfully: (obj: ResultProps) => void;
+}
