+++ src/screens/WalletScreen/Modals/SendModal/screens/index.ts
@@ -1,0 +1,6 @@
+import TransactionContacts from './TransactionContacts';
+import TransactionSelectFunds from './TransactionSelectFunds';
+import TransactionTransfer from './TransactionTransfer';
+import AddContact from './AddContact';
+
+export { TransactionContacts, TransactionSelectFunds, TransactionTransfer, AddContact };
