+++ src/screens/WalletScreen/Modals/SendModal/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	header: {
+		height: 64,
+		flexDirection: 'row',
+		alignItems: 'center',
+		justifyContent: 'space-between',
+		paddingHorizontal: 32
+	}
+});
