+++ src/screens/WalletScreen/Modals/index.ts
@@ -1,0 +1,7 @@
+import SendModal from './SendModal';
+import ComingSoonModal from './ComingSoonModal';
+import ReceiveModal from './ReceiveModal';
+import SentModal from './SentModal';
+import NotAbleToSaveModal from './NotAbleToSaveModal';
+
+export { SendModal, ComingSoonModal, ReceiveModal, SentModal, NotAbleToSaveModal };
