+++ src/screens/WalletScreen/WalletScreen.types.ts
@@ -1,0 +1,6 @@
+type ResultProps = {
+	link: string;
+	symbol: string;
+};
+
+export type { ResultProps };
